fixed more firebase typecast issues and added singleton to firebase

+++ C++urry/WaterNet/app/src/main/java/com/cs2340/WaterNet/Model/Contaminant.java
@@ -1,21 +1,21 @@
 package com.cs2340.WaterNet.Model;
 
 /**
- * Created by pulki on 2/27/2017.
+ * Created by Pulkit Gupta on 2/27/2017.
  */
 
 public enum Contaminant {
     CONTAMINANT("Contaminant", 0);
     String value;
-    int ppm;
+    long ppm;
     Contaminant(String value, int ppm) {
         this.value = value;
         this.ppm = ppm;
     }
-    void setPPM(int ppm) {
+    void setPPM(long ppm) {
         this.ppm = ppm;
     }
-    int getPPM() {
+    long getPPM() {
         return ppm;
     }
     String getValue() {
