+++ C++urry/WaterNet/app/src/main/java/com/cs2340/WaterNet/Model/Virus.java
@@ -1,21 +1,21 @@
 package com.cs2340.WaterNet.Model;
 
 /**
- * Created by pulki on 2/27/2017.
+ * Created by Pulkit Gupta on 2/27/2017.
  */
 
 public enum Virus {
     VIRUS("Contaminant", 0);
     String value;
-    int ppm;
-    Virus(String value, int ppm) {
+    long ppm;
+    Virus(String value, long ppm) {
         this.value = value;
         this.ppm = ppm;
     }
-    void setPPM(int ppm) {
+    void setPPM(long ppm) {
         this.ppm = ppm;
     }
-    int getPPM() {
+    long getPPM() {
         return ppm;
     }
     String getValue() {
